django_backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.32.3
pillow==10.3.0